+++ CHANGES.rst
@@ -24,6 +24,7 @@ in-progress
 - Add page about :ref:`mqttwarn-alerts`
 - Add some research information, mainly around LoRa and related technologies
 - Some words about software watchdog for :ref:`data-loss events`
+- Enable :ref:`Data acquisition over HTTP <daq-http>` and provide an example for :ref:`daq-php`
 
 
 .. _version-0.4.1: